Working through broken tests
